requirements.txt
django-cms-articles
leprikon
